added multiaccounting, filters implemented support for multiaccounting implemented filters for handlers removed default msgtype from instant_messaging types refactored types module to separate response and request types added primitive types to improve typing

+++ aiomatrix/utils/storage/internal_data/engines/sqlite.py
@@ -5,26 +5,27 @@ from ...base_engines import SqliteConnection
 
 
 class SqliteInternalDataStorageEngine(SqliteConnection):
-    async def get_data_by_key(self, key: str) -> Optional[models.InternalDataPair]:
-        sql = 'SELECT * FROM internal_data WHERE key = ?'
-        params = (key,)
+    async def get_data_by_key(self, account_id: str, key: str) -> Optional[models.InternalDataPair]:
+        sql = 'SELECT * FROM internal_data WHERE key = ? AND account_id = ?'
+        params = (key, account_id)
         return self._make_request(sql, params, fetch=True, model_type=models.InternalDataPair)
 
-    async def get_all_data(self) -> List[models.InternalDataPair]:
-        sql = 'SELECT * FROM internal_data WHERE key = ?'
-        return self._make_request(sql, params=None, fetch=True, mult=True, model_type=models.InternalDataPair)
+    async def get_all_data(self, account_id: str) -> List[models.InternalDataPair]:
+        sql = 'SELECT * FROM internal_data WHERE account_id = ?'
+        params = (account_id,)
+        return self._make_request(sql, params=params, fetch=True, mult=True, model_type=models.InternalDataPair)
 
-    async def insert_data(self, key: str, value: str):
-        sql = 'INSERT INTO internal_data (key, data) VALUES (?, ?)'
-        params = (key, f'{value}')
+    async def insert_data(self, account_id: str, key: str, value: str):
+        sql = 'INSERT INTO internal_data (key, data, account_id) VALUES (?, ?, ?)'
+        params = (key, f'{value}', account_id)
         self._make_request(sql, params)
 
-    async def set_data_by_key(self, key: str, value: str):
-        sql = 'UPDATE internal_data SET data = ? WHERE key = ?'
-        params = (f'{value}', key)
+    async def set_data_by_key(self, account_id: str, key: str, value: str):
+        sql = 'UPDATE internal_data SET data = ? WHERE key = ?AND account_id = ?'
+        params = (f'{value}', key, account_id)
         self._make_request(sql, params)
 
-    async def delete_data_by_key(self, key: str):
-        sql = 'DELETE FROM internal_data WHERE key = ?'
-        params = (key,)
+    async def delete_data_by_key(self, account_id: str, key: str):
+        sql = 'DELETE FROM internal_data WHERE key = ? AND account_id = ?'
+        params = (key, account_id)
         self._make_request(sql, params)
